requirements.txt
opencv-python==4.12.0.88
roboflow==1.2.11
supervision==0.26.1
ultralytics==8.3.225
pyserial==3.5
